+++ UMCPClient/Assets/UMCP/Editor/UMCPBridge.cs
@@ -416,6 +416,7 @@ namespace UMCP.Editor
                     "execute_menu_item" => ExecuteMenuItem.HandleCommand(paramsObject),
                     "get_project_path" => GetProjectPath.HandleCommand(paramsObject),
                     "get_unity_state" => GetCurrentState(), // New command for getting current state
+                    "force_update_editor" => ForceUpdateEditor.HandleCommand(paramsObject),
                     "mark_start_of_new_step" => MarkStartOfNewStep.HandleCommand(paramsObject),
                     "request_step_logs" => RequestStepLogs.HandleCommand(paramsObject),
                     _ => throw new ArgumentException($"Unknown or unsupported command type: {command.type}")